fix ansible-pull handling of repo_overrides.yml
Sadly ansible prefetches the vars_files for all the plays before running.
Since we create repo_overrides.yml with play 0 and then use it in vars_files
in the following plays this was a problem.

Workaround: prefetching does not happen if vars_files is a list.
From there comes NOT_EXISTING_HACK